[1822] Small PR review changes.
